requirements.txt
Jinja2==3.1.2 
MarkupSafe==2.1.3 
Werkzeug==2.3.6 
blinker==1.6.2 
click==8.1.4 
colorama==0.4.6 
flask==2.3.2 
itsdangerous==2.1.2
dnspython==2.3.0 
pymongo==4.4.0
python-dotenv==1.0.0
bidict==0.22.1
eventlet==0.33.3 
flask-socketio==5.3.4 
greenlet==2.0.2 
python-engineio==4.5.1 
python-socketio==5.8.0 
six==1.16.0
aiohttp==3.8.4 
aiosignal==1.3.1 
async-timeout==4.0.2 
attrs==23.1.0 
certifi==2023.5.7 
charset-normalizer==3.2.0 
frozenlist==1.3.3 
idna==3.4 
multidict==6.0.4 
openai==0.27.8 
requests==2.31.0 
tqdm==4.65.0 
urllib3==2.0.3 
yarl==1.9.2
flask-cors==4.0.0